# chore(SharedMemory): refactor and small perf improvement