## Changes
- `current_len` is redundant -> replaced with `last_checkpoint` with little performance boost
- some renaming and refactoring done, the code is a little cleaner (imo)
- added a couple of tests
##  Performance

Result of running `cargo bench --all` again main:
```text
analysis/transact/raw   time:   [7.1716 µs 7.2122 µs 7.3133 µs]
                        change: [-9.4337% -6.1147% -2.2493%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
analysis/transact/checked
                        time:   [7.1405 µs 7.1533 µs 7.1852 µs]
                        change: [-9.1607% -7.6801% -6.1826%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
analysis/transact/analysed
                        time:   [5.7375 µs 5.7428 µs 5.7535 µs]
                        change: [-17.280% -10.357% -4.4181%] (p = 0.01 < 0.05)
                        Performance has improved.

snailtracer/transact/analysed
                        time:   [57.387 ms 57.414 ms 57.443 ms]
                        change: [-5.6044% -4.7900% -4.0188%] (p = 0.00 < 0.05)
                        Performance has improved.
snailtracer/eval        time:   [53.667 ms 53.834 ms 54.265 ms]
                        change: [-15.078% -11.167% -7.9253%] (p = 0.00 < 0.05)
                        Performance has improved.

transfer/transact/analysed
                        time:   [1.3286 µs 1.3429 µs 1.3629 µs]
                        change: [+9.3643% +10.574% +11.840%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```